+++ eslint.config.js
@@ -1,7 +1,7 @@
 import typescriptEslint from "@typescript-eslint/eslint-plugin";
 import stylisticTs from '@stylistic/eslint-plugin-ts'
 import jsdoc from "eslint-plugin-jsdoc";
-import jest from "eslint-plugin-jest";
+import vitest from "eslint-plugin-vitest";
 import globals from "globals";
 import tsParser from "@typescript-eslint/parser";
 
@@ -14,7 +14,7 @@ export default [{
         "@typescript-eslint": typescriptEslint,
         "@stylistic": stylisticTs,
         jsdoc,
-        jest,
+        vitest,
     },
 
     linterOptions: {
@@ -108,22 +108,20 @@ export default [{
         "jsdoc/require-returns-description": "warn",
         "jsdoc/check-alignment": "error",
         "jsdoc/check-line-alignment": "error",
-        "jest/no-commented-out-tests": "error",
-        "jest/no-disabled-tests": "warn",
-        "jest/no-focused-tests": "error",
-        "jest/prefer-to-contain": "warn",
-        "jest/prefer-to-have-length": "warn",
-        "jest/valid-expect": "error",
-        "jest/prefer-to-be": "warn",
-        "jest/no-alias-methods": "warn",
-        "jest/no-interpolation-in-snapshots": "warn",
+        "vitest/no-commented-out-tests": "error",
+        "vitest/no-disabled-tests": "warn",
+        "vitest/no-focused-tests": "error",
+        "vitest/prefer-to-contain": "warn",
+        "vitest/prefer-to-have-length": "warn",
+        "vitest/valid-expect": "error",
+        "vitest/prefer-to-be": "warn",
+        "vitest/no-alias-methods": "warn",
+        "vitest/no-interpolation-in-snapshots": "warn",
 
-        "jest/no-large-snapshots": ["warn", {
+        "vitest/no-large-snapshots": ["warn", {
             maxSize: 50,
             inlineMaxSize: 20,
-        }],
-
-        "jest/no-deprecated-functions": "warn",
+        }]
     },
 }, {
     files: ["test/**"],