+++ src/expression/index.ts
@@ -38,6 +38,7 @@ import {NumberArray} from './types/number_array';
 import {ColorArray} from './types/color_array';
 import {VariableAnchorOffsetCollection} from './types/variable_anchor_offset_collection';
 import {ProjectionDefinition} from './types/projection_definition';
+import {GlobalState} from './definitions/global_state';
 
 export type Feature = {
     readonly type: 0 | 1 | 2 | 3 | 'Unknown' | 'Point' | 'MultiPoint' | 'LineString' | 'MultiLineString' | 'Polygon' | 'MultiPolygon';
@@ -166,12 +167,14 @@ export function createExpression(expression: unknown, propertySpec?: StyleProper
 export class ZoomConstantExpression<Kind extends EvaluationKind> {
     kind: Kind;
     isStateDependent: boolean;
+    globalStateRefs: Set<string>;
     _styleExpression: StyleExpression;
 
     constructor(kind: Kind, expression: StyleExpression) {
         this.kind = kind;
         this._styleExpression = expression;
         this.isStateDependent = kind !== ('constant' as EvaluationKind) && !isStateConstant(expression.expression);
+        this.globalStateRefs = findGlobalStateRefs(expression.expression);
     }
 
     evaluateWithoutErrorHandling(
@@ -201,7 +204,7 @@ export class ZoomDependentExpression<Kind extends EvaluationKind> {
     kind: Kind;
     zoomStops: Array<number>;
     isStateDependent: boolean;
-
+    globalStateRefs: Set<string>;
     _styleExpression: StyleExpression;
     interpolationType: InterpolationType;
 
@@ -210,6 +213,7 @@ export class ZoomDependentExpression<Kind extends EvaluationKind> {
         this.zoomStops = zoomStops;
         this._styleExpression = expression;
         this.isStateDependent = kind !== ('camera' as EvaluationKind) && !isStateConstant(expression.expression);
+        this.globalStateRefs = findGlobalStateRefs(expression.expression);
         this.interpolationType = interpolationType;
     }
 
@@ -250,6 +254,7 @@ export function isZoomExpression(expression: any): expression is ZoomConstantExp
 
 export type ConstantExpression = {
     kind: 'constant';
+    globalStateRefs: Set<string>;
     readonly evaluate: (
         globals: GlobalProperties,
         feature?: Feature,
@@ -262,6 +267,7 @@ export type ConstantExpression = {
 export type SourceExpression = {
     kind: 'source';
     isStateDependent: boolean;
+    globalStateRefs: Set<string>;
     readonly evaluate: (
         globals: GlobalProperties,
         feature?: Feature,
@@ -274,6 +280,7 @@ export type SourceExpression = {
 
 export type CameraExpression = {
     kind: 'camera';
+    globalStateRefs: Set<string>;
     readonly evaluate: (
         globals: GlobalProperties,
         feature?: Feature,
@@ -289,6 +296,7 @@ export type CameraExpression = {
 export type CompositeExpression = {
     kind: 'composite';
     isStateDependent: boolean;
+    globalStateRefs: Set<string>;
     readonly evaluate: (
         globals: GlobalProperties,
         feature?: Feature,
@@ -407,6 +415,7 @@ export function normalizePropertyExpression<T>(
             constant = ProjectionDefinition.parse(value);
         }
         return {
+            globalStateRefs: new Set<string>(),
             kind: 'constant',
             evaluate: () => constant
         };
@@ -452,6 +461,17 @@ function findZoomCurve(expression: Expression): Step | Interpolate | ExpressionP
     });
 
     return result;
+}
+
+export function findGlobalStateRefs(expression: Expression, results = new Set<string>()): Set<string> {
+    if (expression instanceof GlobalState) {
+        results.add(expression.key);
+    }
+
+    expression.eachChild(childExpression => {
+        findGlobalStateRefs(childExpression, results);
+    });
+    return results;
 }
 
 function getExpectedType(spec: StylePropertySpecification): Type {