+++ CHANGELOG.md
@@ -1,12 +1,19 @@
 ## main
 
 ### ✨ Features and improvements
-- Restructure expression syntax documentation ([#1112](https://github.com/maplibre/maplibre-style-spec/pull/1112))
 - Add new layer type: `elevation` ([#1067](https://github.com/maplibre/maplibre-style-spec/issues/1067))
 - _...Add new stuff here..._
 
 ### 🐞 Bug fixes
+- Add missing `global-state` expression typing ([#1130](https://github.com/maplibre/maplibre-style-spec/pull/1130))
 - _...Add new stuff here..._
+
+## 23.2.3
+
+### ✨ Features and improvements
+
+- Restructure expression syntax documentation ([#1112](https://github.com/maplibre/maplibre-style-spec/pull/1112))
+- Add global state ref support ([#1120](https://github.com/maplibre/maplibre-style-spec/pull/1120))
 
 ## 23.2.2